Refactor: introduce ele_new function
Make comments and the code structure more readable.
Pack up the repeated code in q_insert_head and q_insert_tail.
In this function, it allocates the list element and the string,
and check if the allocations are valid.
It returns a boolean value to the insertion function, representing
that the allocations are success or not.