Revise: implement q_remove_head
Directly return false if the queue is NULL or empty.
Use strdup to copy the string if needed.
Remove the head element from the list. A different operation is
needed if there is only one element in the list.
Lastly, free the space used by the list element and the string.