Revise: q_remove_head do not store removed value
The parameter passed to this function is a pointer to character,
so using strndup to allocate new space and assign it to the pointer
won't change the real value which is originally pointed by the
pointer.
So, we use strncpy instead. This will directly copy the removed
value to the space pointed by the pointer. More, we need to
manually assign the terminated character to the end of string.