BugFix: tail ptr points to wrong ele after sort
The problem:
The tail pointer is not handled in the sorting function, so it
will still point to the last element in the queue before sorting.

The solution:
Pass the pointer to the tail pointer to the function, use the
local variables left_end and right_end to record the pointer to the
tail of the partitioned lists, and assign the right_end to the tail
pointer after sorting.