Revise: introduce two new variables to merge sort
The problem:
If the two lists now merging are already sorted, which means the
remaining elements in one are always smaller than ones in the other,
there will be unnecessary traverse because the merging process will
always traverse through the whole two lists.

The solution:
Use a_tail, b_tail to specify the tail of the two lists now sorting.
If the situation mentioned above occurs, there will be only one
pointer changed, which can improve the efficiency in some cases.